Apply bouncer denials in a deterministic order
It's possible (likely, even!) for a route to include multiple bouncers
that deny. For example, one bouncer could deny with an authentication
error, while another denies with an authorization error. We want both
denials to have the ability to modify the request with a middleware
method, but the denials also have the ability to interrupt the
processing chain and immediately respond. For this reason, we want to
make sure that bouncers that deny first (meaning they were defined first
in the code) get the first opportunity to respond. In our example of a
request with both an authentication and an authorization error, the
authentication middleware might want to respond immediately, in which
case the authorization middleware should not first get the opportunity
to modify the request.

(paired w/ Max.)