handle uncaught exceptions in bouncer middleware
Because we weren't try/catching the call and a synchronous error was being thrown
in async code (`async.eachSeries`), eventually the response would hang untli the
client closed the connection. Unintuitively, the response code would be 200 because
it was never over-written.

Additionally, now if a bouncer middleware itself throws a synchronous error, this
will also be handled. In both of these cases a 500 is now returned immediately.

We discovered this due to some of our apps using multiple versions of secure-router
(one in goodeggs-server and one directly). This lead to bouncers being setup with
an old pattern that didn't include a `middleware` function to return the error.

pair w/ @bobzoller